simplify and abstract gathering context for upload
