+++ cli-tests/src/test.rs
@@ -1,7 +1,11 @@
+use std::{fs, io::BufReader};
+
 use assert_cmd::Command;
+use assert_matches::assert_matches;
+use bundle::BundleMeta;
 use predicates::prelude::*;
 use tempfile::tempdir;
-use test_utils::mock_server::MockServerBuilder;
+use test_utils::mock_server::{MockServerBuilder, RequestPayload};
 
 use crate::utils::{
     generate_mock_codeowners, generate_mock_git_repo, generate_mock_valid_junit_xmls, CARGO_RUN,
@@ -112,6 +116,34 @@ async fn test_command_fails_with_no_junit_files_no_quarantine_successful_upload(
         ));
 
     println!("{assert}");
+
+    let requests = state.requests.lock().unwrap().clone();
+    assert_eq!(requests.len(), 4);
+    let mut requests_iter = requests.into_iter();
+
+    assert!(matches!(
+        requests_iter.next().unwrap(),
+        RequestPayload::CreateRepo(..)
+    ));
+    assert!(matches!(
+        requests_iter.next().unwrap(),
+        RequestPayload::CreateBundleUpload(..)
+    ));
+
+    let tar_extract_directory =
+        assert_matches!(requests_iter.next().unwrap(), RequestPayload::S3Upload(d) => d);
+    let file = fs::File::open(tar_extract_directory.join("meta.json")).unwrap();
+    let reader = BufReader::new(file);
+    let bundle_meta: BundleMeta = serde_json::from_reader(reader).unwrap();
+    assert_eq!(
+        bundle_meta.base_props.test_command.unwrap(),
+        "bash -c exit 128"
+    );
+
+    assert!(matches!(
+        requests_iter.next().unwrap(),
+        RequestPayload::UpdateBundleUpload(..)
+    ));
 }
 
 #[tokio::test(flavor = "multi_thread")]
