+++ cli/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod api_client;
+pub mod context;
 pub mod print;
 pub mod quarantine;
-pub mod scanner;
 pub mod test;
 pub mod upload;
 pub mod validate;
